composables/useValidation: extract validation fns
